requirements.txt
streamlit
speechrecognition
python-dotenv
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-groq
faiss-cpu
sentence-transformers
gtts
pyttsx3
pyaudio